Improvement
No need for hasClass helper - read this: 
https://developer.mozilla.org/en-US/docs/Web/API/Element/classList

No need for check on your line 15 - nobody is using window event. 
No need to cache e.target - you're using it only once.

For If condition it's better to invert it and return false if it's not true - you have less indents in this case.